# DLTE-PROJECT/api-docs.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8082","description":"Generated server url"}],"paths":{"/profile/register":{"post":{"tags":["my-bank-officials-api"],"operationId":"save","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MyBankOfficials"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MyBankOfficials"}}}}}}},"/pay/add":{"post":{"tags":["my-controller"],"summary":"This is used for adding new Payee","operationId":"newPayee","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Payee"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"409":{"description":"Payee account and sender's account cannot be same","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"Payee Inserted successfully","content":{"*/*":{"schema":{"type":"string"}}}},"204":{"description":"No record exists","content":{"*/*":{"schema":{"type":"string"}}}},"403":{"description":"Sender account is inactive","content":{"*/*":{"schema":{"type":"string"}}}},"500":{"description":"Internal server error","content":{"*/*":{"schema":{"type":"string"}}}}}}}},"components":{"schemas":{"GrantedAuthority":{"type":"object","properties":{"authority":{"type":"string"}}},"MyBankOfficials":{"type":"object","properties":{"userId":{"type":"integer","format":"int64"},"name":{"type":"string"},"userName":{"type":"string"},"password":{"type":"string"},"email":{"type":"string"},"status":{"type":"integer","format":"int32"},"attempts":{"type":"integer","format":"int32"},"maxAttempt":{"type":"integer","format":"int32"},"username":{"type":"string"},"enabled":{"type":"boolean"},"authorities":{"type":"array","items":{"$ref":"#/components/schemas/GrantedAuthority"}},"accountNonExpired":{"type":"boolean"},"accountNonLocked":{"type":"boolean"},"credentialsNonExpired":{"type":"boolean"}}},"Payee":{"required":["payeeAccountNumber","payeeName","senderAccountNumber"],"type":"object","properties":{"payeeId":{"type":"integer","format":"int32"},"senderAccountNumber":{"type":"integer","format":"int64"},"payeeAccountNumber":{"type":"integer","format":"int64"},"payeeName":{"pattern":"^[a-zA-Z\\s]*$","type":"string"}}}}}}